fix: change bottom bar buttons
With the changes introduced between FLEX 3 and 5,
there is now a bottom bar with new options. Since we
can't really take advantage of them due to our feature
stripping, we should hide these options.

The size of the icons was increased to make better use
of the space of the bar.

References https://outsystemsrd.atlassian.net/browse/RNMT-5718